+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Accept folders to open and navigate the content in the new directory navigation sidebar.
 - [AppStore ver.] Now the user can donate to the CotEditor project via in-app purchase in the new Donate settings pane.
 - Add new “Select Enclosing Symbols” and “Split Selection by Lines” commands to the Edit > Select menu.
+- Add new “Show in Finder” command to the File menu.
 - Add new “Create Document in CotEditor” shortcut action.
 - Support the alpha channel for the current line in theme settings.
 - Add Assembly syntax.